build(electron): move keyboard-config.model to shared-with-frontend to fix ASAR require
