+++ docs/wiki/3.03-Keyboard-Shortcuts.md
@@ -139,7 +139,7 @@ Specifically for the **Next Month** shortcut, the task is scheduled for the **1s
 
 **Where shortcuts are defined:**
 
-- Config model: `KeyboardConfig` (keyboard-config.model.ts).
+- Config model: `KeyboardConfig` (electron/shared-with-frontend/keyboard-config.model.ts).
 - Default mappings: `default-global-config.const.ts` (keyboard section).
 - Settings form (which actions appear and their labels): `keyboard-form.const.ts`.
