+++ electron/shared-with-frontend/keyboard-config.model.ts
